crazy-waymo: retire the night-windows pass, keep the mast luminaires
NightWindows painted lit windows onto the kit buildings. The kits are gone
and the parcel fabric lights its own windows, so the pass detected nothing
("0 lit windows on 0 glazed + 0 procedural buildings") and only its
side-job survived: the luminaire on downtown's signal masts. That moves
to fx/street-luminaires.ts as a plain registration at world-loader's
playable gate; the 1.1k-line pass, its per-frame intensity hook and the
WorldCoreSystems slot go.

Same-camera night pair at the Financial District: mean luminance 95.5 vs
95.4, hot share 20.41% both ways.

Closes #315
